refactor: Extract scenario data
